Relocate Teleop folder
Also renamed Miscellaneous folder to correct spelling
