+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Miscellaneous/ConceptScanServo.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.Miscelaneous;
+package org.firstinspires.ftc.teamcode.Miscellaneous;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
